requirements.txt
tensorflow-gpu
seaborn